feat(results): generate results page
When completing the quiz, the app will redirect to /results. The query params correlate to the top 3
gift keys for the user's quiz results. For example 0=a&1=b&2=c corresponds to gifts [a, b, c] as the
user's top three gifts. This URL is short and bookmarkable, if a user should choose to do that.